lsp : サンプリング周波数の引き回しをuint32_tからfloatへ変更

+++ libsynthpp/src/lsp/effector/envelope_generator.hpp
@@ -90,7 +90,7 @@ public:
 
 	// エンベロープ形状パラメータを指定します (メロディ用)
 	void setMelodyEnvelope(
-		uint32_t sampleFreq,		    // Hz
+		float sampleFreq,				// Hz
 		float attack_time,				// sec
 		float decay_time,				// sec
 		parameter_type sustain_level,	// level (0 <= x <= 1)
@@ -100,7 +100,7 @@ public:
 		setMelodyEnvelope(sampleFreq, {}, attack_time, 0, decay_time, sustain_level, 0, release_time, std::numeric_limits<parameter_type>::epsilon());
 	}
 	void setMelodyEnvelope(
-		uint32_t sampleFreq,			// Hz
+		float sampleFreq,				// Hz
 		Curve curve,					
 		float attack_time,				// sec
 		float hold_time,				// sec
@@ -125,7 +125,7 @@ public:
 	}
 	// エンベロープ形状パラメータを指定します (ドラム用)
 	void setDrumEnvelope(
-		uint32_t sampleFreq,			// Hz
+		float sampleFreq,				// Hz
 		Curve curve,
 		float attack_time,				// sec
 		float hold_time,				// sec
